project/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
torchvision==0.17.2+cpu
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
